Fix broken substream cache if stream was aborted by the user
When a channel ends we have to cache the substream based on how much
data we actually recived, not based on the ideal channel length.

Otherwise we end up caching the last few seconds of a stream aborted
by the user as a full channel length, and the next time the track is
played we get cached data which is incomplete and we wrongly assume
the track is done.